finalize structuring feed pages
